feat: Add skipRounding flag into EV [DHIS2-15689]
