+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/org/hisp/dhis/eventvisualization/hibernate/EventVisualization.hbm.xml
@@ -141,6 +141,8 @@
 
         <property name="completedOnly" column="completedonly"/>
 
+        <property name="skipRounding"/>
+
         <property name="timeField" column="timefield"/>
 
         <property name="orgUnitField" column="orgunitfield"/>
