+++ dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.41/V2_41_17__Add_sorting_to_eventvisualization.sql
@@ -1,4 +1,4 @@
 -- https://dhis2.atlassian.net/browse/DHIS2-14956
 -- Adds a new column "sorting" into EventVisualization table.
 
-alter table eventvisualization add column if not exists "sorting" jsonb default '[]'::jsonb;
+alter table eventvisualization add column if not exists "skip" boolean;
