+++ dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.41/V2_41_27__Add_skip_rounding_to_eventvisualization.sql
@@ -1,0 +1,4 @@
+-- https://dhis2.atlassian.net/browse/DHIS2-15689
+-- Adds a new column "skiprounding" into EventVisualization table.
+
+alter table eventvisualization add column if not exists "skiprounding" boolean;
